Speed up and simplify describeCommand(). (dm)
git-svn-id: svn://mielke.cc/main/brltty@4555 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/cmd.h
@@ -40,7 +40,7 @@ typedef struct {
 extern const CommandEntry commandTable[];
 extern const CommandEntry *getCommandEntry (int code);
 
-extern void describeCommand (int command, char *buffer, int size);
+extern void describeCommand (int command, char *buffer, size_t size);
 
 extern int enqueueCommand (int command);
 extern int dequeueCommand (void);
